refactor: split fs folder for more sub folder
