+++ src/commandsList.js
@@ -1,5 +1,4 @@
 import {handleFsOperations} from './fs/handleFsOperations.js';
-import {up} from './fs/up.js';
 
 export const commandsMap = new Map([
     ['.exit', process.exit],
